backend/requirements.txt
flask
flask-cors
requests
python-dotenv
openai
google-generativeai
